pyproject.toml
[project]
name = "pythonstudy"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "numpy",
    "opencv-python",
    "pandas",
    "paramiko",
    "pillow",
    "psycopg2-binary",
    "sqlalchemy",
]
